# public/openapi.yaml
openapi: 3.0.1
info:
  title: TODO Plugin
  description: A plugin that allows the user to create and manage a TODO list using ChatGPT.
  version: 'v1'
servers:
  - url: http://localhost:3000
paths:
...
  /api/eth-gpt:
    post:
      operationId: resolveName
      summary: Resolve the ENS name to address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/resolveNameRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resolveNameResponse'
...
components:
  schemas:
...
    resolveNameRequest:
      type: object
      properties:
        ensName:
          type: string
          description: The ENS name we are trying to resolve.
    resolveNameResponse:
      type: object
      properties:
        address:
          type: string
          description: The Ethereum address of that requested ENS name resolves to.